Cache relative paths instead of using an angular filter.
This will improve the performance a lot. First load performance may
suffer a bit, though.

Also, the leading "./" is skipped now, as this causes additional string
ops, which turn out to be particularly harmful in some cases (slots vs
heaps aka. shortstrings in spidermonkey, allocation overhead). And the
leading "./" is not really required anymore.