Fast fix for broken filters and/or quick access stuff
ng-if will create a new scope, so any `set` operations are on that new
scope and not reflected in the parent scope.
Using ng-show instead will not create a new scope and therefore does not
have this issue. But different from ng-if, ng-show still creates a DOM
but just hides it...