Added support for bind IP in the CLI params.
In addition to the first command line argument supporting a custom bind port (from the default 8888),
added support for a custom bind IP as the second command line argument.

Useful when the user may want to run the server on a particular network interface and not all (which is the default behaviour).

The code change keeps the default behaviour as is.

Signed-off-by: Md Safiyat Reza <safiyat@voereir.com>